fix(core): address Copilot review on PR #65
Extracted the [assembly: TranspileAssembly] detection (semantic +
syntactic fallback) into a shared SymbolHelper.HasTranspileAssembly
extension. Both the IR producer (CSharpSourceFrontend) and the legacy
TypeTransformer.TransformAll now call it, so the two paths can no
longer drift.

Part of #58
